requirements.txt
jax
dm-haiku
matplotlib
numpy
scipy
pandas